Remove sections of readme better served in docs
If the readme is "marketing" for people looking to see if they want to
install the plugin, then the vimdoc is a "manual" to help them figure
out how to use it once installed.

These sections duplicate what's in the vimdoc and are more relevant to
someone trying to configure the plugin instead of someone checking out
what it can do.

Also, these sections don't have images.